refactor: ui -> view package로 이동
