+++ src/main/kotlin/racing/view/InputView.kt
@@ -1,4 +1,4 @@
-package racing.ui
+package racing.view
 
 import racing.domain.CarName
