+++ src/main/kotlin/racing/view/OutputView.kt
@@ -1,4 +1,4 @@
-package racing.ui
+package racing.view
 
 import racing.domain.Car
 import racing.domain.Cars
